% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academicqualtrics-package.R
\docType{package}
\name{academicqualtrics-package}
\alias{academicqualtrics}
\alias{academicqualtrics-package}
\title{academicqualtrics: Access the Qualtrics API Endpoint}
\description{
A wrapper for Qualtrics API
}
\author{
\strong{Maintainer}: Justin Chun-ting Ho \email{j.c.ho@uva.nl}

}
\keyword{internal}
