% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{create_block_randomizer}
\alias{create_block_randomizer}
\title{Create Randomizer Block}
\usage{
create_block_randomizer(
  blockids,
  DATA_CENTER,
  SURVEY_ID,
  API_TOKEN,
  present = 5,
  even = TRUE
)
}
\arguments{
\item{blockids}{character vector, ids of text blocks}

\item{DATA_CENTER}{string, your Qualtrics data center}

\item{SURVEY_ID}{string, your Qualtrics survey ID}

\item{API_TOKEN}{string, your Qualtrics API token}

\item{present}{integer, how many blocks to show to each annotator}

\item{even}{boolean, should the randomizer distribute text evenly}
}
\description{
Create Randomizer Block
}
\examples{
blockids <- c("BL_v7iQbRtKtAv7mQ10zO1h", "BL_t4VtQCrLDxzJalrgShVx")
\dontrun{
create_block_randomizer(blockids,
                        "fra1",
                        "SV_S3A96bzOnfyKMEDCiKhw",
                        "ZAhIjt6CkPO5FyczlRhJ")
}
}
