% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_survey}
\alias{get_survey}
\title{Title}
\usage{
get_survey(DATA_CENTER, SURVEY_ID, API_TOKEN)
}
\arguments{
\item{DATA_CENTER}{string, your Qualtrics data center}

\item{SURVEY_ID}{string, your Qualtrics survey ID}

\item{API_TOKEN}{string, your Qualtrics API token}
}
\value{
json, survey setup
}
\description{
Title
}
\examples{
\dontrun{
get_survey(             "fra1",
                        "SV_S3A96bzOnfyKMEDCiKhw",
                        "ZAhIjt6CkPO5FyczlRhJ")
}
}
