% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolific.R
\name{setup_prolific}
\alias{setup_prolific}
\title{Setup Prolific Integration
This function replicates the steps specified here: https://researcher-help.prolific.com/en/article/fbbd36}
\usage{
setup_prolific(url, DATA_CENTER, SURVEY_ID, API_TOKEN)
}
\arguments{
\item{url}{string, url for the Prolific return URL}

\item{DATA_CENTER}{string, your Qualtrics data center}

\item{SURVEY_ID}{string, your Qualtrics survey ID}

\item{API_TOKEN}{string, your Qualtrics API token}
}
\description{
Setup Prolific Integration
This function replicates the steps specified here: https://researcher-help.prolific.com/en/article/fbbd36
}
\examples{
blockids <- c("BL_v7iQbRtKtAv7mQ10zO1h", "BL_t4VtQCrLDxzJalrgShVx")
question <- c("What does a Hobbit hole look like?")
answers <- c("nasty","dirty","comfortable")
\dontrun{
setup_prolific("https://app.prolific.com/submissions/complete?cc=A20A223T",
              "fra1",
              "SV_S3A96bzOnfyKMEDCiKhw",
              "ZAhIjt6CkPO5FyczlRhJ")
}
}
