% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{add_questions}
\alias{add_questions}
\title{Add Questions to Text Blocks}
\usage{
add_questions(
  blockids,
  question,
  answers,
  tags,
  DATA_CENTER,
  SURVEY_ID,
  API_TOKEN,
  selector = "SAVR",
  subselector = "TX",
  forced = "ON"
)
}
\arguments{
\item{blockids}{character vector, ids of text blocks}

\item{question}{string, question for annotators}

\item{answers}{character vector, answers for annotators}

\item{tags}{character vector, identifier for texts, needs to be same length as x}

\item{DATA_CENTER}{string, your Qualtrics data center}

\item{SURVEY_ID}{string, your Qualtrics survey ID}

\item{API_TOKEN}{string, your Qualtrics API token}

\item{selector}{string, the format used for collecting responses, one of: "DL", "GRB", "MACOL", "MAHR", "MAVR", "MSB", "NPS", "SACOL", "SAHR", "SAVR", "SB", "TB", "TXOT", "PTB"}

\item{subselector}{string, additional options or variations for response collection, one of: "GR", "TX", "TXOT", "WOTXB", "WTXB"}

\item{forced}{string, force respondents to answer a question, one of: "ON", "OFF"}
}
\description{
Add Questions to Text Blocks
}
\examples{
blockids <- c("BL_v7iQbRtKtAv7mQ10zO1h", "BL_t4VtQCrLDxzJalrgShVx")
question <- c("What does a Hobbit hole look like?")
answers <- c("nasty","dirty","comfortable")
\dontrun{
add_questions(blockids,
              question,
              answers,
              tags,
              "fra1",
              "SV_S3A96bzOnfyKMEDCiKhw",
              "ZAhIjt6CkPO5FyczlRhJ")
}
}
