% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{add_questions}
\alias{add_questions}
\title{Add Questions to Text Blocks}
\usage{
add_questions(blockids, question, answers, DATA_CENTER, SURVEY_ID, API_TOKEN)
}
\arguments{
\item{blockids}{character vector, ids of text blocks}

\item{question}{string, question for annotators}

\item{answers}{character vector, answers for annotators}

\item{DATA_CENTER}{string, your Qualtrics data center}

\item{SURVEY_ID}{string, your Qualtrics survey ID}

\item{API_TOKEN}{string, your Qualtrics API token}
}
\description{
Add Questions to Text Blocks
}
\examples{
blockids <- c("BL_v7iQbRtKtAv7mQ10zO1h", "BL_t4VtQCrLDxzJalrgShVx")
question <- c("What does a Hobbit hole look like?")
answers <- c("nasty","dirty","comfortable")
\dontrun{
add_questions(blockids,
              question,
              answers,
              "fra1",
              "SV_S3A96bzOnfyKMEDCiKhw",
              "ZAhIjt6CkPO5FyczlRhJ")
}
}
