% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{add_texts}
\alias{add_texts}
\title{Create Text Blocks for Annotation}
\usage{
add_texts(x, DATA_CENTER, SURVEY_ID, API_TOKEN)
}
\arguments{
\item{x}{character vector, texts to be annotated}

\item{DATA_CENTER}{string, your Qualtrics data center}

\item{SURVEY_ID}{string, your Qualtrics survey ID}

\item{API_TOKEN}{string, your Qualtrics API token}
}
\value{
A character vector of block ids. It is necessary for later steps.
}
\description{
Create Text Blocks for Annotation
}
\examples{
texts <- c("In a hole in the ground there lived a hobbit.",
            "It was a hobbit-hole, and that means comfort.")
\dontrun{
add_texts(texts,
          "fra1",
          "SV_S3A96bzOnfyKMEDCiKhw",
          "ZAhIjt6CkPO5FyczlRhJ")
}
}
